+++ src/components/MonitorView.js
@@ -1,63 +1,55 @@
 import React, { Component } from 'react';
-import './App.css';
 import NumericSignal from '../components/NumericSignal'
 import GraphSignal from '../components/GraphSignal';
-import Settings from '../utils/Settings';
-import SettingsModal from '../utils/SettingsModal';
 import contrastIcon from '../images/contrastIcon.png';
 const EVENT_STREAM_ADDRESS = "http://localhost:5000/events"; //    http://141.45.146.235:8200
 
-
-
 class MonitorView extends Component {
-
-    constructor(props) {
-      super(props);
-      this.evtSource = new EventSource(EVENT_STREAM_ADDRESS);
-      this.state = {
-        connection: "not connected",
-        ekg: {min: 20, max: 80, current: 0},
-        pulse: {min: 20, max: 80, current: 0},
-        temperature: {min: 20, max: 80, current: 0},
-        oxygen: {min: 20, max: 80, current: 0},
-        heartrate: {min: 20, max: 80, current: 0},
-        systole: {min: 20, max: 80, current: 0},
-        diastole: {min: 20, max: 80, current: 0},
-        showDarkContrast: false,
-      };
-    }
-
-    componentDidMount() {
-    var that = this;
-
-    this.evtSource.onerror = function (e) {
-      console.log("error: ", e);
-      that.setState(Object.assign({}, { connection: "error" }));
+  constructor(props) {
+    super(props);
+    this.evtSource = new EventSource(EVENT_STREAM_ADDRESS);
+    this.state = {
+      connection: "not connected",
+      ekg: { min: 20, max: 80, current: 0 },
+      pulse: { min: 20, max: 80, current: 0 },
+      temperature: { min: 20, max: 80, current: 0 },
+      oxygen: { min: 20, max: 80, current: 0 },
+      heartrate: { min: 20, max: 80, current: 0 },
+      systole: { min: 20, max: 80, current: 0 },
+      diastole: { min: 20, max: 80, current: 0 },
     };
-
-    this.evtSource.onopen = function (e) {
-      console.log("open: ", e);
-      while(e.srcElement.readyState == 0) {
-        console.log("connecting...");
-        that.setState(Object.assign({}, { connection: "connecting..." }));
-      }
-      that.setState(Object.assign({}, { connection: "connected" }));
-    };
-
-    this.evtSource.close = function (e) {
-      console.log("close: ",this.evtSource.readyState);
-      that.setState(Object.assign({}, { connection: "not connected" }));
-    };
-
-
-    this.subscribeSSE("ekg");
-    this.subscribeSSE("pulse");
-    this.subscribeSSE("temperature");
-    this.subscribeSSE("heartrate");
-    this.subscribeSSE("oxygen");
-    this.subscribeSSE("diastole");
-    this.subscribeSSE("systole");
   }
+
+  componentDidMount() {
+  var that = this;
+
+  this.evtSource.onerror = function (e) {
+    console.log("error: ", e);
+    that.setState(Object.assign({}, { connection: "error" }));
+  };
+
+  this.evtSource.onopen = function (e) {
+    console.log("open: ", e);
+    while(e.srcElement.readyState == 0) {
+      console.log("connecting...");
+      that.setState(Object.assign({}, { connection: "connecting..." }));
+    }
+    that.setState(Object.assign({}, { connection: "connected" }));
+  };
+
+  this.evtSource.close = function (e) {
+    console.log("close: ",this.evtSource.readyState);
+    that.setState(Object.assign({}, { connection: "not connected" }));
+  };
+
+  this.subscribeSSE("ekg");
+  this.subscribeSSE("pulse");
+  this.subscribeSSE("temperature");
+  this.subscribeSSE("heartrate");
+  this.subscribeSSE("oxygen");
+  this.subscribeSSE("diastole");
+  this.subscribeSSE("systole");
+}
 
 // This function subscribes a server side event for each vital sign to which it is applied:
   subscribeSSE(vitalSign) {
@@ -66,8 +58,6 @@ class MonitorView extends Component {
       that.setState({[vitalSign]: Object.assign({...that.state[vitalSign]}, {current: e.data})});
     }, false);
   }
-
-
 
   componentDidUpdate() {
     if(this.state.readyState == 0){
@@ -82,20 +72,18 @@ class MonitorView extends Component {
   }
 
   isCritical(vitalSign) {
-    return this.state[vitalSign].current<this.state[vitalSign].min || this.state[vitalSign].current>this.state[vitalSign].max;
+    return this.state[vitalSign].current < this.state[vitalSign].min ||
+      this.state[vitalSign].current > this.state[vitalSign].max;
   }
 
   render() {
-
     return (
       <div className="app-container">
-      <SettingsModal />
         <div id="header">
           OKTOPUS BIOSIGNAL MONITORING
         </div>
 
         <div id="grid-container">
-
             <GraphSignal
             className="biosignal-EKG"
             bioSignalType="EKG"
@@ -121,7 +109,6 @@ class MonitorView extends Component {
             <NumericSignal bioSignalType="HR / min" warning={this.isCritical('heartrate')} bioSignalValue={this.state.heartrate.current} className="biosignal-HR" />
             <NumericSignal bioSignalType="spO2 %" warning={this.isCritical('oxygen')} bioSignalValue={this.state.oxygen.current} className="biosignal-sO2"/>
             <NumericSignal bioSignalType="Sys / Dia mmHG" warning={this.isCritical('diastole') || this.isCritical('systole')} bioSignalValue={this.state.diastole.current} bioSignalValueAddOn={this.state.systole.current} className="biosignal-RR"/>
-
         </div>
 
         <div id="footer">connection status: {this.state.connection}.</div>
