--- src/container/MenuView.js
@@ -1,21 +1,0 @@
-import React, { Component } from 'react';
-import './App.css';
-import MonitorView from './MonitorView.js';
-
-
-
-class MenuView extends Component {
-
-    constructor(props) {
-      super(props);
-    }
-
-  render() {
-
-    return (
-    <p>Menu</p>
-    );
-  }
-}
-
-export default MenuView;
