--- src/index.css
@@ -1,4 +1,0 @@
-
-body {
-    font-family: 'Archivo Black', sans-serif;
-  }
