+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom'
-import App from '../src/container/App';
-import './index.css'
+import App from '../src/components/App';
+import './App.css';
 
 // Since we are using HtmlWebpackPlugin WITHOUT a template, we should create our own root node in the body element before rendering into it
 let root = document.createElement('div')
